requirements.txt
pip == 23.3.1
Django == 4.2.6
djangorestframework == 3.14.0
psycopg2 == 2.9.9
django-crispy-forms == 1.14.0
crispy-bootstrap5 == 0.7
django-environ==0.4.5
djoser==2.2.0
django-cors-headers==3.7.0
pyJWT==2.8.0
social-auth-app-django==5.4.0
whitenoise==6.6.0
argon2-cffi==23.1.0
django-ckeditor == 6.7.0
pytz == 2023.3
Pillow == 10.1.0
mutagen == 1.47.0
stripe == 7.6.0
braintree == 4.24.0
boto3 == 1.29.5
django-storages == 1.14.2
markdown == 3.5.1
gunicorn == 21.2.0
requests == 2.31.0
django-ckeditor == 6.7.0
Pillow == 10.1.0


